backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# Reddit (using public JSON API via httpx)

# Sentiment analysis
vaderSentiment==3.3.2

# LLM
google-generativeai==0.8.3

# Report generation
python-pptx==1.0.2
python-docx==1.1.2
jinja2==3.1.4

# Auth
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Payments
stripe==8.0.0

# Utils
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
email-validator==2.1.0
gunicorn==22.0.0
